+++ src/Configuration/Configuration.php
@@ -21,6 +21,8 @@ use Fidry\FileSystem\FS;
 use Humbug\PhpScoper\Configuration\Configuration as PhpScoperConfiguration;
 use Humbug\PhpScoper\Container;
 use InvalidArgumentException;
+use KevinGH\Box\Annotation\CompactedFormatter;
+use KevinGH\Box\Annotation\DocblockAnnotationParser;
 use KevinGH\Box\Compactor\Compactor;
 use KevinGH\Box\Compactor\Compactors;
 use KevinGH\Box\Compactor\Php as PhpCompactor;
@@ -34,6 +36,7 @@ use KevinGH\Box\Phar\CompressionAlgorithm;
 use KevinGH\Box\Phar\SigningAlgorithm;
 use KevinGH\Box\PhpScoper\SerializableScoper;
 use Phar;
+use phpDocumentor\Reflection\DocBlockFactory;
 use RuntimeException;
 use Seld\JsonLint\ParsingException;
 use SplFileInfo;
@@ -1667,24 +1670,22 @@ final class Configuration
         ?array $ignoredAnnotations,
         ConfigurationLogger $logger,
     ): array {
-        return array_filter(
-            array_map(
-                static function (string $class) use ($raw, $basePath, $logger, $ignoredAnnotations): ?Compactor {
-                    Assert::classExists($class, 'The compactor class %s does not exist.');
-                    Assert::isAOf($class, Compactor::class, sprintf('The class "%s" is not a compactor class.', $class));
+        return array_map(
+            static function (string $class) use ($raw, $basePath, $logger, $ignoredAnnotations): Compactor {
+                Assert::classExists($class, 'The compactor class %s does not exist.');
+                Assert::isAOf($class, Compactor::class, sprintf('The class "%s" is not a compactor class.', $class));
 
-                    if (in_array($class, [PhpCompactor::class, 'KevinGH\Box\Compactor\Php'], true)) {
-                        return self::createPhpCompactor($ignoredAnnotations);
-                    }
+                if (in_array($class, [PhpCompactor::class, 'KevinGH\Box\Compactor\Php'], true)) {
+                    return self::createPhpCompactor($ignoredAnnotations);
+                }
 
-                    if (in_array($class, [PhpScoperCompactor::class, 'KevinGH\Box\Compactor\PhpScoper'], true)) {
-                        return self::createPhpScoperCompactor($raw, $basePath, $logger);
-                    }
+                if (in_array($class, [PhpScoperCompactor::class, 'KevinGH\Box\Compactor\PhpScoper'], true)) {
+                    return self::createPhpScoperCompactor($raw, $basePath, $logger);
+                }
 
-                    return new $class();
-                },
-                $compactorClasses,
-            ),
+                return new $class();
+            },
+            $compactorClasses,
         );
     }
 
@@ -2670,10 +2671,10 @@ final class Configuration
         return $ignored;
     }
 
-    private static function createPhpCompactor(?array $ignoredAnnotations): ?Compactor
+    private static function createPhpCompactor(?array $ignoredAnnotations): Compactor
     {
         if (null === $ignoredAnnotations) {
-            return null;
+            return new PhpCompactor(null);
         }
 
         $ignoredAnnotations = array_values(
@@ -2685,7 +2686,13 @@ final class Configuration
             ),
         );
 
-        return PhpCompactor::create($ignoredAnnotations);
+        return new PhpCompactor(
+            new DocblockAnnotationParser(
+                DocBlockFactory::createInstance(),
+                new CompactedFormatter(),
+                $ignoredAnnotations,
+            ),
+        );
     }
 
     private static function createPhpScoperCompactor(
