+++ src/Console/Application.php
@@ -15,7 +15,7 @@ declare(strict_types=1);
 namespace KevinGH\Box\Console;
 
 use Fidry\Console\Application\Application as FidryApplication;
-use KevinGH\Box\Parallel\ProcessFileCommand;
+use KevinGH\Box\Parallelization\ProcessFileCommand;
 use function KevinGH\Box\get_box_version;
 use function sprintf;
 use function trim;
