+++ src/Parallelization/Configuration.php
@@ -12,18 +12,19 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace KevinGH\Box\Parallel;
+namespace KevinGH\Box\Parallelization;
 
-use Humbug\PhpScoper\Symbol\SymbolsRegistry;
+use KevinGH\Box\Compactor\Compactors;
+use KevinGH\Box\MapFile;
+use function serialize;
+use function unserialize;
 
-final readonly class BatchResult
+final readonly class Configuration
 {
-    /**
-     * @param array{string, string} $processedFilesWithContents
-     */
     public function __construct(
-        public array $processedFilesWithContents,
-        public SymbolsRegistry $symbolsRegistry,
+        public array $filePaths,
+        public MapFile $mapFile,
+        public Compactors $compactors,
     ) {
     }
 
@@ -38,19 +39,5 @@ final readonly class BatchResult
     public static function unserialize(string $serialized): self
     {
         return unserialize($serialized);
-    }
-
-    /**
-     * @return array{
-     *     array{string, string},
-     *     SymbolsRegistry,
-     * }
-     */
-    public function toArray(): array
-    {
-        return [
-            $this->processedFilesWithContents,
-            $this->symbolsRegistry,
-        ];
     }
 }
