+++ src/Parallelization/ProcessFileCommand.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace KevinGH\Box\Parallel;
+namespace KevinGH\Box\Parallelization;
 
 use Fidry\FileSystem\FS;
 use Humbug\PhpScoper\Symbol\SymbolsRegistry;
